requirements.txt
duckdb
pandas
pyarrow
requests
pyyaml
